Add support for AGP 9
This adds a convenience compat class that supports backwards compatibility as well as support for LibraryExtension, which will he required with AGP9.

Upgrading to AGP 9 will also require extension developers to replace BaseExtension in the `Project.android` declaration with LibraryExtension, and also replace `compileSdkVersion` with `compileSdk` outside of `defaultConfig`, as well as move `targetSdk` to a new lint block, and remove `apply(plugin = "kotlin-android")` since AGP 9 has built-in Kotlin, that is no longer supported.

This should have no affecf for those that don't upgrade to AGP 9, or those you don't decide to use LibraryExtension in AGP 8 which is also supported.